query cost and handling multiple graphs
